+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Changes
 
 - Add `elasticstack_fleet_output` data source. ([#1762](https://github.com/elastic/terraform-provider-elasticstack/pull/1762))
+- Fix `termField` validation for ESQL `.es-query` alert rules in `elasticstack_kibana_alerting_rule`. ([#1914](https://github.com/elastic/terraform-provider-elasticstack/pull/1914))
 
 ## [0.14.3] - 2026-03-02